tests: fix unexpected error from put object via grpc_utils
Sometimes the request landed on a node that doesn't belong
to the container, it caused a non related error message:
"local operation on the node not compliant with the container storage policy".
While test expects "strorage group type is no longer supported".
