+++ pytest_tests/tests/object/test_object_put_grpc.py
@@ -1,5 +1,8 @@
+import logging
+import random
+
 import allure
-from helpers.container import create_container
+from helpers.container import create_container, parse_container_nodes_output
 from helpers.file_helper import generate_file
 from helpers.grpc_utils import put_object
 from helpers.neofs_verbs import put_object as put_object_via_cli
@@ -7,6 +10,8 @@ from helpers.wellknown_acl import ALLOW_ALL_OPERATIONS_EXCEPT_DELETE
 from neofs_testlib.env.env import NeoFSEnv, NodeWallet
 from neofs_testlib.protobuf.generated.object import types_pb2 as object_types_pb2
 from neofs_testlib.protobuf.generated.session import types_pb2 as session_types_pb2
+
+logger = logging.getLogger("NeoLogger")
 
 
 def test_put_storage_group_object_no_longer_supported(default_wallet: NodeWallet, neofs_env: NeoFSEnv):
@@ -17,8 +22,18 @@ def test_put_storage_group_object_no_longer_supported(default_wallet: NodeWallet
             endpoint=neofs_env.sn_rpc,
             rule="REP 1",
         )
+
     with allure.step("Try to put a storage group object via grpc"):
-        response = put_object(object_types_pb2.ObjectType.STORAGE_GROUP, default_wallet, neofs_env, b"payload", cid)
+        nodes = parse_container_nodes_output(
+            neofs_env.neofs_cli(neofs_env.storage_nodes[0].cli_config)
+            .container.nodes(rpc_endpoint=neofs_env.storage_nodes[0].endpoint, cid=cid)
+            .stdout
+        )
+
+        node = random.choice(nodes)
+        node_endpoint = f"localhost:{node['endpoint'].split('/')[-1]}"
+
+        response = put_object(object_types_pb2.ObjectType.STORAGE_GROUP, default_wallet, node_endpoint, b"payload", cid)
         assert response.meta_header.status.message == "strorage group type is no longer supported"
 
 
@@ -41,7 +56,7 @@ def test_put_tombstone_object_without_delete_permission(
         response = put_object(
             object_types_pb2.ObjectType.TOMBSTONE,
             not_owner_wallet,
-            neofs_env,
+            neofs_env.storage_nodes[0].endpoint,
             b"payload",
             cid,
             oid,
@@ -53,7 +68,7 @@ def test_put_tombstone_object_without_delete_permission(
         response = put_object(
             object_types_pb2.ObjectType.TOMBSTONE,
             not_owner_wallet,
-            neofs_env,
+            neofs_env.storage_nodes[0].endpoint,
             b"payload",
             cid,
             oid,
@@ -82,7 +97,12 @@ def test_put_object_without_homo_hash(default_wallet: NodeWallet, neofs_env_sing
 
     with allure.step("Try to put an object via grpc without homo hash"):
         response = put_object(
-            object_types_pb2.ObjectType.REGULAR, default_wallet, neofs_env_single_sn, b"payload", cid, homo_hash=False
+            object_types_pb2.ObjectType.REGULAR,
+            default_wallet,
+            neofs_env_single_sn.storage_nodes[0].endpoint,
+            b"payload",
+            cid,
+            homo_hash=False,
         )
         assert response.meta_header.status.message != "missing homomorphic payload checksum"
 
@@ -105,6 +125,11 @@ def test_put_object_without_homo_hash(default_wallet: NodeWallet, neofs_env_sing
 
     with allure.step("Try to put an object via grpc without homo hash"):
         response = put_object(
-            object_types_pb2.ObjectType.REGULAR, default_wallet, neofs_env_single_sn, b"payload", cid, homo_hash=False
+            object_types_pb2.ObjectType.REGULAR,
+            default_wallet,
+            neofs_env_single_sn.storage_nodes[0].endpoint,
+            b"payload",
+            cid,
+            homo_hash=False,
         )
         assert response.meta_header.status.message == "missing homomorphic payload checksum"
